@@ -25,48 +25,43 @@ public class ExportGUI
         final int result = fc.showOpenDialog(FilePanel);
 
         final Thread[] exporthread = {null};
-        exporthread[0] = new Thread(new Runnable()
-        {
-            @Override
-            public void run()
+        exporthread[0] = new Thread(() -> {
+
+            if (result == JFileChooser.APPROVE_OPTION)
             {
+                File chosen = fc.getSelectedFile().isDirectory() ? fc.getSelectedFile() : fc.getSelectedFile().getParentFile();
+                Date date = new Date(System.currentTimeMillis());
+                SimpleDateFormat df2 = new SimpleDateFormat("MM-dd-yy, HH.mm.ss");
+                String dateText = df2.format(date);
 
-                if (result == JFileChooser.APPROVE_OPTION)
+                File log = new File(chosen, "Lava Control @ " + dateText + ".log");
+
+                try
                 {
-                    File chosen = fc.getSelectedFile().isDirectory() ? fc.getSelectedFile() : fc.getSelectedFile().getParentFile();
-                    Date date = new Date(System.currentTimeMillis());
-                    SimpleDateFormat df2 = new SimpleDateFormat("MM-dd-yy, HH.mm.ss");
-                    String dateText = df2.format(date);
-
-                    File log = new File(chosen, "Lava Control @ " + dateText + ".log");
-
-                    try
+                    BufferedWriter bw = new BufferedWriter(new FileWriter(log));
+                    for (String line : Main.consoleGUI.getConsoleText().getText().split("\n"))
                     {
-                        BufferedWriter bw = new BufferedWriter(new FileWriter(log));
-                        for (String line : Main.consoleGUI.getConsoleText().getText().split("\n"))
-                        {
-                            bw.write(line + "\n");
-                        }
-
-                        bw.flush();
-                        bw.close();
-                    }
-                    catch (IOException e)
-                    {
-                        e.printStackTrace();
+                        bw.write(line + "\n");
                     }
 
-                    JOptionPane.showMessageDialog(Main.consoleFrame, "The log \"LavaControl_" + dateText + "\" has been successfully saved!");
+                    bw.flush();
+                    bw.close();
+                }
+                catch (IOException e)
+                {
+                    e.printStackTrace();
+                }
 
-                    try
-                    {
-                        exporthread[0].join();
-                    }
-                    catch (InterruptedException e)
-                    {
-                        exporthread[0].interrupt();
-                        //e.printStackTrace();
-                    }
+                JOptionPane.showMessageDialog(Main.consoleFrame, "The log \"LavaControl_" + dateText + "\" has been successfully saved!");
+
+                try
+                {
+                    exporthread[0].join();
+                }
+                catch (InterruptedException e)
+                {
+                    exporthread[0].interrupt();
+                    //e.printStackTrace();
                 }
             }
         });
