@@ -1,18 +1,17 @@
 package com.colonelhedgehog.lavacontrol.core;
 
+import com.colonelhedgehog.lavacontrol.core.components.JPlaceholderTextField;
 import com.colonelhedgehog.lavacontrol.core.components.RoundedCornerBorder;
 
 import javax.swing.*;
-import java.awt.Image;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
+import java.awt.*;
 
 /**
  * Created by ColonelHedgehog on 8/19/15.
  */
 public class PrefsGUI
 {
-    private JPanel PrefsPanel;
+    private JPanel prefsPanel;
     private JTextField memoryBash;
     private JTextField maxConsoleLines;
     private JSeparator serverSSeperator;
@@ -28,9 +27,15 @@ public class PrefsGUI
     private JLabel passwordLabel;
     private JTabbedPane tabbedPane;
     private JCheckBox askToExport;
+    private JCheckBox saveJarToPathCheckBox;
+    private JCheckBox generateDocsCheckBox;
+    private JCheckBox generateSourcesCheckBox;
+    private JPlaceholderTextField versionField;
 
     public PrefsGUI()
     {
+        versionField.setDefaultText("latest");
+
         final Settings settings = Main.getSettings();
         settings.reloadSettings();
 
@@ -40,11 +45,17 @@ public class PrefsGUI
         sshUsername.setText(settings.getSSHUsername());
         sshPassword.setText(settings.getSSHPassword());
         sshHost.setText(settings.getSSHHost());
+
         stickyScrollbar.setSelected(settings.getStickyScrollBar());
         closeOnStop.setSelected(settings.getCloseWindowOnStop());
         askToExport.setSelected(settings.getAskExportLog());
 
-        if(closeOnStop.isSelected())
+        saveJarToPathCheckBox.setSelected(settings.saveJarToPath());
+        generateDocsCheckBox.setSelected(settings.generateDocs());
+        generateSourcesCheckBox.setSelected(settings.generateSources());
+        versionField.setText(settings.getBuildVersion());
+
+        if (closeOnStop.isSelected())
         {
             askToExport.setSelected(false);
             askToExport.setEnabled(false);
@@ -56,122 +67,112 @@ public class PrefsGUI
 
         disableSSH();
 
-        sshEnabled.addActionListener(new ActionListener()
-        {
-            @Override
-            public void actionPerformed(ActionEvent e)
+        sshEnabled.addActionListener(e -> {
+            sshUsername.setEnabled(sshEnabled.isSelected());
+            sshHost.setEnabled(sshEnabled.isSelected());
+            sshPassword.setEnabled(sshEnabled.isSelected());
+            passwordLabel.setEnabled(sshEnabled.isSelected());
+        });
+
+        closeOnStop.addActionListener(e -> {
+            if (closeOnStop.isSelected())
             {
-                sshUsername.setEnabled(sshEnabled.isSelected());
-                sshHost.setEnabled(sshEnabled.isSelected());
-                sshPassword.setEnabled(sshEnabled.isSelected());
-                passwordLabel.setEnabled(sshEnabled.isSelected());
+                askToExport.setSelected(false);
+                askToExport.setEnabled(false);
+            }
+            else
+            {
+                askToExport.setEnabled(true);
             }
         });
 
-        closeOnStop.addActionListener(new ActionListener()
-        {
-            @Override
-            public void actionPerformed(ActionEvent e)
+        applyButton.addActionListener(e -> {
+            String errors = "";
+            try
             {
-                if(closeOnStop.isSelected())
-                {
-                    askToExport.setSelected(false);
-                    askToExport.setEnabled(false);
-                }
-                else
-                {
-                    askToExport.setEnabled(true);
-                }
+                settings.setMemBash(Integer.parseInt(memoryBash.getText()));
             }
+            catch (NumberFormatException nfe)
+            {
+                String message = Main.Prefix + "ERROR: Text in memory bash settings is invalid.";
+                System.out.println(message);
+                errors = errors + "\n" + message;
+            }
+
+            try
+            {
+                settings.setMaxConsoleLines(Integer.parseInt(maxConsoleLines.getText()));
+            }
+            catch (NumberFormatException nfe)
+            {
+                String message = Main.Prefix + "ERROR: Text in max console lines settings is\ninvalid.";
+                System.out.println(message);
+                errors = errors + "\n" + message;
+            }
+
+            settings.setSSHEnabled(sshEnabled.isSelected());
+
+            settings.setSSHUsername(sshUsername.getText());
+            settings.setSSHPassword(getText(sshPassword.getPassword()));
+            settings.setSSHHost(sshHost.getText());
+
+            settings.setStickyScrollBar(stickyScrollbar.isSelected());
+            settings.setCloseWindowOnStop(closeOnStop.isSelected());
+            settings.setAskExportLog(askToExport.isSelected());
+
+            settings.setSaveJarToPath(saveJarToPathCheckBox.isSelected());
+            settings.setGenerateDocs(generateDocsCheckBox.isSelected());
+            settings.setGenerateSources(generateSourcesCheckBox.isSelected());
+            settings.setBuildVersion(versionField.getText());
+
+            settings.saveSettings();
+
+            ImageIcon scaled = new ImageIcon(Main.getIcon().getImage().getScaledInstance(96, 96, Image.SCALE_SMOOTH));
+            if (!errors.equals(""))
+            {
+                JOptionPane.showMessageDialog(Main.prefsFrame, "<html><b>Error!</b> Couldn't save your settings:\n" + errors + "\n\nPlease fix these errors and then hit apply again\nto completely save all settings.", "", JOptionPane.ERROR_MESSAGE, scaled);
+            }
+            else
+            {
+                JOptionPane.showMessageDialog(Main.prefsFrame, "<html><b>Success!</b></html> Settings were successfully applied.", "", JOptionPane.INFORMATION_MESSAGE, scaled);
+            }
+
         });
 
-        applyButton.addActionListener(new ActionListener()
-        {
-            @SuppressWarnings("deprecation")
-            @Override
-            public void actionPerformed(ActionEvent e)
-            {
-                String errors = "";
-                try
-                {
-                    settings.setMemBash(Integer.parseInt(memoryBash.getText()));
-                }
-                catch(NumberFormatException nfe)
-                {
-                    String message = "[Lava Control] ERROR: Text in memory bash settings is invalid.";
-                    System.out.println(message);
-                    errors = errors + "\n" + message;
-                }
+        restoreDefaultsButton.addActionListener(e -> {
+            Main.getSettings().deleteSettings();
 
-                try
-                {
-                    settings.setMaxConsoleLines(Integer.parseInt(maxConsoleLines.getText()));
-                }
-                catch(NumberFormatException nfe)
-                {
-                    String message = "[Lava Control] ERROR: Text in max console lines settings is\ninvalid.";
-                    System.out.println(message);
-                    errors = errors + "\n" + message;
-                }
-
-                settings.setSSHEnabled(sshEnabled.isSelected());
-
-                settings.setSSHUsername(sshUsername.getText());
-                settings.setSSHPassword(sshPassword.getText());
-                settings.setSSHHost(sshHost.getText());
-
-                settings.setStickyScrollBar(stickyScrollbar.isSelected());
-                settings.setCloseWindowOnStop(closeOnStop.isSelected());
-                settings.setAskExportLog(askToExport.isSelected());
-
-                settings.saveSettings();
-
-                ImageIcon scaled = new ImageIcon(Main.getIcon().getImage().getScaledInstance(96, 96, Image.SCALE_SMOOTH));
-                if(!errors.equals(""))
-                {
-                    JOptionPane.showMessageDialog(Main.prefsFrame, "<html><b>Error!</b> Couldn't save your settings:\n" + errors + "\n\nPlease fix these errors and then hit apply again\nto completely save all settings.", "", JOptionPane.ERROR_MESSAGE, scaled);
-                }
-                else
-                {
-                    JOptionPane.showMessageDialog(Main.prefsFrame, "<html><b>Success!</b></html> Settings were successfully applied.", "", JOptionPane.INFORMATION_MESSAGE, scaled);
-                }
-
-            }
+            memoryBash.setText(String.valueOf(settings.getMemBash()));
+            maxConsoleLines.setText(String.valueOf(settings.getMaxConsoleLines()));
+            sshEnabled.setSelected(settings.getSSHEnabled());
+            sshUsername.setText(settings.getSSHUsername());
+            sshPassword.setText(settings.getSSHPassword());
+            sshHost.setText(settings.getSSHHost());
+            stickyScrollbar.setSelected(settings.getStickyScrollBar());
+            closeOnStop.setSelected(settings.getCloseWindowOnStop());
+            askToExport.setSelected(settings.getAskExportLog());
         });
 
-        restoreDefaultsButton.addActionListener(new ActionListener()
-        {
-            @Override
-            public void actionPerformed(ActionEvent e)
-            {
-                Main.getSettings().deleteSettings();
-
-                memoryBash.setText(String.valueOf(settings.getMemBash()));
-                maxConsoleLines.setText(String.valueOf(settings.getMaxConsoleLines()));
-                sshEnabled.setSelected(settings.getSSHEnabled());
-                sshUsername.setText(settings.getSSHUsername());
-                sshPassword.setText(settings.getSSHPassword());
-                sshHost.setText(settings.getSSHHost());
-                stickyScrollbar.setSelected(settings.getStickyScrollBar());
-                closeOnStop.setSelected(settings.getCloseWindowOnStop());
-                askToExport.setSelected(settings.getAskExportLog());
-            }
-        });
-
-        cancelButton.addActionListener(new ActionListener()
-        {
-            @Override
-            public void actionPerformed(ActionEvent e)
-            {
-                Main.prefsFrame.setVisible(false);
-            }
-        });
+        cancelButton.addActionListener(e -> Main.prefsFrame.setVisible(false));
 
         memoryBash.setBorder(new RoundedCornerBorder(6));
         sshUsername.setBorder(new RoundedCornerBorder(6));
         sshHost.setBorder(new RoundedCornerBorder(6));
         sshPassword.setBorder(new RoundedCornerBorder(6));
         maxConsoleLines.setBorder(new RoundedCornerBorder(6));
+        versionField.setBorder(new RoundedCornerBorder(6));
+    }
+
+    private String getText(char[] password)
+    {
+        String string = "";
+
+        for (char cha : password)
+        {
+            string += cha;
+        }
+
+        return string;
     }
 
     private void disableSSH()
@@ -183,7 +184,7 @@ public class PrefsGUI
 
     public JPanel getPrefsPanel()
     {
-        return PrefsPanel;
+        return prefsPanel;
     }
 
     public JTextField getMemoryBash()
